requirements.txt
captum == 0.8.0 
kagglehub == 0.3.8
matplotlib == 3.10.1
numpy == 2.2.5
opencv == 4.11.0
pandas == 2.2.3
python == 3.12.10
pytorch == 2.4.0
scikit-learn == 1.6.1
scipy == 1.15.2
seaborn == 0.13.2
torchvision == 0.19.0
tqdm == 4.67.1